requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
executing==2.1.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.10.0
git-filter-repo==2.45.0
huggingface-hub==0.26.2
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jiggler==0.0.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
npm==0.1.1
numpy==2.1.1
optional-django==0.1.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.3
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pynput==1.7.7
pyobjc-core==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
tabulate==0.9.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tokenizers==0.20.3
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
wcwidth==0.2.13
redis==5.0.1
django-redis==5.4.0
